:wrench: release_manager and checkpoint impl to check latests release and infor user
