+++ shell/shell_darwin.go
@@ -1,0 +1,9 @@
+package shell
+
+func GetShell() string {
+	return "bash"
+}
+
+func GetDeploymentPath() (string, error) {
+	return "/usr/local/bin/tlm", nil
+}
