+++ shell/shell_linux.go
@@ -1,0 +1,5 @@
+package shell
+
+func GetShell() string {
+	return "bash"
+}
