+++ shell/shell_windows.go
@@ -1,0 +1,5 @@
+package shell
+
+func GetShell() string {
+	return "powershell"
+}
